Lower video buffers count for 64MB T20.
All the example code in the SDK specify nVBs = 3, but the compiled binary will
not run. VBMCreatePool will always fail with "IMP_ALLOC failed size = 9400320"
when trying to call IMP_FrameSource_EnableChn. Binaries shipped with the
Xiaofang 1S use a value of 2, and this seems to be the fix.

Additionally, undocumented xxx_SetPoolSize() functions need to be called.
These were also found in the iCamera binary.